Use different isolinux.cfg on CentOS-7

+++ scripts/build_netinstall.sh
@@ -203,50 +203,16 @@ perfSONAR Toolkit    Integrated by the perfSONAR Team  Build Date:
 http://www.perfsonar.net  Hit enter to continue    $BUILD_DATE
 EOF
 
-cat > $TEMP_DIRECTORY/isolinux/isolinux.cfg <<EOF
-default vesamenu.c32
-#prompt 1
-timeout 600
+sed -e "s/\\[BUILD_VERSION\\]/$BUILD_VERSION/" \
+    -e "s/\\[KS_FILE\\]/$BUILD_OS_LOWER-$BUILD_TYPE_LOWER.cfg/" \
+    $SCRIPTS_DIRECTORY/../isolinux/centos$BUILD_OS_VERSION_MAJOR.cfg > $TEMP_DIRECTORY/isolinux/isolinux.cfg
 
-display boot.msg
-
-menu background splash.jpg
-menu title Welcome to perfSONAR Toolkit $BUILD_VERSION!
-menu color border 0 #ffffffff #00000000
-menu color sel 7 #ffffffff #ff000000
-menu color title 0 #ffffffff #00000000
-menu color tabmsg 0 #ffffffff #00000000
-menu color unsel 0 #ffffffff #00000000
-menu color hotsel 0 #ff000000 #ffffffff
-menu color hotkey 7 #ffffffff #ff000000
-menu color scrollbar 0 #ffffffff #00000000
-
-label linux
-  menu label ^Install the perfSONAR Toolkit
-  menu default
-  kernel vmlinuz
-  append initrd=initrd.img ks=file:///$BUILD_OS_LOWER-$BUILD_TYPE_LOWER.cfg
-label vesa
-  menu label Install the perfSONAR Toolkit in text mode
-  kernel vmlinuz
-  append initrd=initrd.img text xdriver=vesa nomodeset ks=file:///$BUILD_OS_LOWER-$BUILD_TYPE_LOWER.cfg
-label rescue
-  menu label ^Rescue installed system
-  kernel vmlinuz
-  append initrd=initrd.img rescue
-label local
-  menu label Boot from ^local drive
-  localboot 0xffff
-label memtest86
-  menu label ^Memory test
-  kernel memtest
-  append -
-EOF
-
-echo "Building boot logo file."
-convert $LOGO_FILE ppm:- | ppmtolss16 '#FFFFFF=7' > $TEMP_DIRECTORY/isolinux/splash.lss
-convert -depth 16 -colors 65536 $LOGO_FILE $TEMP_DIRECTORY/isolinux/splash.png
-mv $TEMP_DIRECTORY/isolinux/splash.png $TEMP_DIRECTORY/isolinux/splash.jpg
+if [ -f $LOGO_FILE ]; then
+	echo "Building boot logo file."
+	convert $LOGO_FILE ppm:- | ppmtolss16 '#FFFFFF=7' > $TEMP_DIRECTORY/isolinux/splash.lss
+	convert -depth 16 -colors 65536 $LOGO_FILE $TEMP_DIRECTORY/isolinux/splash.png
+	mv $TEMP_DIRECTORY/isolinux/splash.png $TEMP_DIRECTORY/isolinux/splash.jpg
+fi
 
 ##############################
 # Create new ISO and MD5 and Cleanup
